+++ src/Controller/SecurityController.php
@@ -38,23 +38,29 @@ class SecurityController extends AbstractController
     }
 
     /**
-    * @Route("/users/profil/modifier", name="app_profil_change")
+    * @Route("/users/profil/modifier/{id}", name="app_profil_change")
     */
-    public function editProfil(Request $request, EntityManagerInterface $em)
+    public function editProfil($id, Request $request, EntityManagerInterface $em)
     {
-        $user = $this->getUser();
-
-        $editProfile_form = $this->createForm(EditProfileType::class, $user);
-        $editProfile_form->handleRequest($request); 
-        
-        if($editProfile_form->isSubmitted() && $editProfile_form->isValid())
+        if($this->isCsrfTokenValid('update'.$id, $request->query->get('csrf')))
         {
-            $em->persist($user);
-            $em->flush();
+            $user = $this->getUser();
 
-            $this->addFlash('update', 'Profil mis à jour');
-            return $this->redirectToRoute('app_profil');
-        }
+            $editProfile_form = $this->createForm(EditProfileType::class, $user);
+            $editProfile_form->handleRequest($request);
+
+            if($editProfile_form->isSubmitted() && $editProfile_form->isValid())
+            {
+                $em->persist($user);
+                $em->flush();
+
+                $this->addFlash('update', 'Profil mis à jour');
+                return $this->redirectToRoute('app_profil');
+            }
+        }else{
+            $this->addFlash('error', 'Votre lien n\'est pas valide !');
+            return $this->redirectToRoute('app_home');
+        } 
 
         return $this->render('security/profilChange.html.twig', [
             'editProfilForm' => $editProfile_form->createView()
@@ -64,27 +70,33 @@ class SecurityController extends AbstractController
     /**
     * @Route("/users/pass/modifier/{id}", name="app_pass_change")
     */
-    public function editPass(Request $request, UserPasswordHasherInterface $userPasswordHasherInterface)
+    public function editPass($id, Request $request, UserPasswordHasherInterface $userPasswordHasherInterface)
     {
-        $user = $this->getUser('id');
-
-        $passEdit_form = $this->createForm(ChangePasswordFormType::class);
-        $passEdit_form->handleRequest($request);
-        
-        if($passEdit_form->isSubmitted() && $passEdit_form->isValid())
+        if($this->isCsrfTokenValid('update'.$id, $request->query->get('csrf')))
         {
-            $encodedPassword = $userPasswordHasherInterface->hashPassword(
-                $user,
-                $passEdit_form->get('plainPassword')->getData()
-            );
-                
-            $user->setPassword($encodedPassword);
-            $this->getDoctrine()->getManager()->flush();
+            $user = $this->getUser('id');
 
-            $this->addFlash('update', 'Votre mot de passe a bien été modifié !');
-            return $this->redirectToRoute('app_profil');
-        }
-           
+            $passEdit_form = $this->createForm(ChangePasswordFormType::class);
+            $passEdit_form->handleRequest($request);
+
+            if($passEdit_form->isSubmitted() && $passEdit_form->isValid())
+            {
+                $encodedPassword = $userPasswordHasherInterface->hashPassword(
+                    $user,
+                    $passEdit_form->get('plainPassword')->getData()
+                );
+
+                $user->setPassword($encodedPassword);
+                $this->getDoctrine()->getManager()->flush();
+
+                $this->addFlash('update', 'Votre mot de passe a bien été modifié !');
+                return $this->redirectToRoute('app_profil');
+            }
+        }else{
+            $this->addFlash('error', 'Votre lien n\'est pas valide !');
+            return $this->redirectToRoute('app_home');
+        } 
+
         return $this->render('security/passChange.html.twig',[
             'editPassForm' => $passEdit_form->createView()
         ]);
@@ -93,17 +105,22 @@ class SecurityController extends AbstractController
     /**
     * @Route("/users/delete/{id}", name="app_pass_delete")
     */
-    public function deletePass(EntityManagerInterface $em, UsersRepository $users, $id)
+    public function deletePass(EntityManagerInterface $em, UsersRepository $users, $id, Request $request)
     {
-        $user = $users->find($id);
-                
-        $session = $this->get('session');
-        $session = new Session();
-        $session->invalidate();
-        $em->remove( $user);
-        $em->flush();
-        return $this->redirectToRoute('app_home');
-        $this->addFlash('delete', 'Votre profil a bien été supprimé !');
+        if($this->isCsrfTokenValid('delete'.$id, $request->query->get('csrf')))
+        {
+            $user = $users->find($id);
+            $session = $this->get('session');
+            $session = new Session();
+            $session->invalidate();
+            $em->remove($user);
+            $em->flush();
+            return $this->redirectToRoute('app_home');
+            $this->addFlash('delete', 'Votre profil a bien été supprimé !');
+        }else{
+            $this->addFlash('error', 'Votre lien n\'est pas valide !');
+            return $this->redirectToRoute('app_home');
+        }
     }
 
     /**
